testing_server/requirements.txt
acme==1.1.0
altgraph==0.17.4
amqp==5.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==5.0.1
backports.zoneinfo==0.2.1
bcrypt==4.2.0
billiard==4.2.1
blinker==1.6.2
build==1.2.2.post1
celery==5.5.2
certbot==0.40.0
certbot-apache==0.39.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.6
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==0.13.0
configobj==5.0.6
coverage==7.6.1
cryptography==40.0.2
dbus-python==1.2.16
distro==1.4.0
Django==4.2.20
django-cors-headers==4.4.0
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.4.0
drf-yasg==1.21.10
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Cors==4.0.0
future==0.18.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
idna==3.10
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
josepy==1.2.0
kombu==5.5.3
MarkupSafe==2.1.3
mock==3.0.5
netaddr==0.8.0
netifaces==0.10.4
numpy==1.24.4
packaging==23.1
paramiko==3.5.0
parsedatetime==2.4
pbr==5.4.5
pillow==10.4.0
pip-tools==7.4.1
pkg_resources==0.0.0
pluggy==1.5.0
prompt_toolkit==3.0.51
psutil==5.9.5
pycairo==1.26.1
pycparser==2.22
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.8
PyJWT==2.9.0
pymongo==4.5.0
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyproject_hooks==1.2.0
pyRFC3339==1.1
pyroute2==0.7.9
pytest==8.3.5
pytest-cov==5.0.0
pytest-django==4.11.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-order==1.3.0
python-augeas==0.5.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2025.2
PyYAML==5.3.1
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.2.1
typing_extensions==4.7.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.3.6
zipp==3.20.2
zope.component==6.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
